in work 10.10.2022

+++ server/src/main/java/ru/explore/with/me/client/event/EventViewStats.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class EventViewStats {
     private String app;
     private String uri;
-    private long hits;
+    private int hits;
 }
